+++ CHANGELOG.md
@@ -6,10 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [25.10] - 2025-04 [Unreleased]
 
+### Added
+
+- current page can be pinned
+
 ### Changed
 
 - uses wex 25.10
-- updated GitHub actions to use wex-version variable
 
 ## [25.04] - 2025-03-16